backend/requirements.txt
boto3>=1.35.76
celery>=5.4,<6
Django>=5.1,<5.2
django-celery-results>=2.5,<3
django-cleanup>=6.0,<7
django-cors-headers>=4.6,<5
django-extensions>=3.2,<4
django-storages>=1.14,<2
django-filter>=24.3,<25
djangorestframework>=3.15,<4
djangorestframework-simplejwt>=5.3,<6
dj-rest-auth>=7.0,<8
pyjwt>=2.9,<3
drf_spectacular>=0.28,<1
optuna>=4.1,<5
pandas>=2.2,<3
irspack==0.4.0
numpy>=2.1,<3
scikit-learn>=1.6,<2
whitenoise>=6.8,<7
django-environ>=0.11,<1
psycopg2-binary>=2.9.10
redis>=5.0,<6
channels>=4.1,<5
channels-redis>=4.2,<5
daphne>=4.1,<5
